[Refactor tests] Separate source and target entities and factories (#72)
