+++ tests/Fixtures/Config/address.yaml
@@ -6,14 +6,14 @@ neusta_converter:
         address: ~
 
 services:
-  Neusta\ConverterBundle\Tests\Fixtures\Model\AddressFactory: ~
+  Neusta\ConverterBundle\Tests\Fixtures\Model\Target\Factory\AddressFactory: ~
   Neusta\ConverterBundle\Tests\Fixtures\Populator\AddressPopulator: ~
 
   test.address.converter:
     public: true
     parent: 'neusta_converter.generic_converter'
     arguments:
-      $factory: '@Neusta\ConverterBundle\Tests\Fixtures\Model\AddressFactory'
+      $factory: '@Neusta\ConverterBundle\Tests\Fixtures\Model\Target\Factory\AddressFactory'
       $populators:
         - '@Neusta\ConverterBundle\Tests\Fixtures\Populator\AddressPopulator'
