+++ tests/Fixtures/Config/person.yaml
@@ -2,7 +2,7 @@ neusta_converter:
   converter:
     test.person.converter:
 #      converter: Neusta\ConverterBundle\Converter\GenericConverter
-      target_factory: Neusta\ConverterBundle\Tests\Fixtures\Model\PersonFactory
+      target_factory: Neusta\ConverterBundle\Tests\Fixtures\Model\Target\Factory\PersonFactory
       populators:
         - Neusta\ConverterBundle\Tests\Fixtures\Populator\PersonNamePopulator
 #      properties:
@@ -13,5 +13,5 @@ neusta_converter:
 #        locale: language # different property names
 
 services:
-  Neusta\ConverterBundle\Tests\Fixtures\Model\PersonFactory: ~
+  Neusta\ConverterBundle\Tests\Fixtures\Model\Target\Factory\PersonFactory: ~
   Neusta\ConverterBundle\Tests\Fixtures\Populator\PersonNamePopulator: ~
