+++ tests/Fixtures/Model/Source/Address.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Neusta\ConverterBundle\Tests\Fixtures\Model;
+namespace Neusta\ConverterBundle\Tests\Fixtures\Model\Source;
 
-class PersonAddress
+class Address
 {
     private string $postalCode;
     private string $city;
