+++ tests/Fixtures/Model/Source/User.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Neusta\ConverterBundle\Tests\Fixtures\Model;
+namespace Neusta\ConverterBundle\Tests\Fixtures\Model\Source;
+
+use Neusta\ConverterBundle\Tests\Fixtures\Model\UnknownType;
 
 class User
 {
