+++ tests/Fixtures/Populator/AddressPopulator.php
@@ -6,8 +6,8 @@ namespace Neusta\ConverterBundle\Tests\Fixtures\Populator;
 
 use Neusta\ConverterBundle\Converter\Context\GenericContext;
 use Neusta\ConverterBundle\Populator;
-use Neusta\ConverterBundle\Tests\Fixtures\Model\Address;
-use Neusta\ConverterBundle\Tests\Fixtures\Model\PersonAddress;
+use Neusta\ConverterBundle\Tests\Fixtures\Model\Source\Address;
+use Neusta\ConverterBundle\Tests\Fixtures\Model\Target\PersonAddress;
 
 /**
  * @implements Populator<Address, PersonAddress, GenericContext>
