+++ tests/Populator/PropertyMappingPopulatorIntegrationTest.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace Neusta\ConverterBundle\Tests\Populator;
 
 use Neusta\ConverterBundle\Tests\ConfigurableKernelTestCase;
-use Neusta\ConverterBundle\Tests\Fixtures\Model\Person;
-use Neusta\ConverterBundle\Tests\Fixtures\Model\User;
+use Neusta\ConverterBundle\Tests\Fixtures\Model\Source\User;
+use Neusta\ConverterBundle\Tests\Fixtures\Model\Target\Person;
 use Neusta\ConverterBundle\Tests\Support\Attribute\ConfigureContainer;
 
 class PropertyMappingPopulatorIntegrationTest extends ConfigurableKernelTestCase
